# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "ExplainRequest": {
        "description": "Body for ``POST /api/explain`` (Iteration 6). ``preferences`` and ``schedule`` are\nthe payload and the result of the solve being explained \u2014 the same\n``SolvePreferences`` sent to ``POST /api/solve`` and the ``SolveResult.schedule`` it\nreturned. Never re-solved: a second solve under the ``-t4`` portfolio could return a\ndifferent equally-optimal plan (ADR-0007), which would explain a schedule the user is\nnot looking at.",
        "properties": {
          "budget_s": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 60.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Overall wall-clock deadline for this explanation (expert-mode setting, mirroring `SolveRequest.time_limit_s`). Omit for the default (`explain.DEFAULT_EXPLAIN_BUDGET_S`). Capped at 60s \u2014 this is a feasibility check, not a search, and the Stage 1 pilot's worst observed batch was 7.6s.",
            "title": "Budget S"
          },
          "preferences": {
            "$ref": "#/components/schemas/SolvePreferences",
            "description": "The edits submitted to the solve being explained."
          },
          "released": {
            "additionalProperties": {
              "type": "integer"
            },
            "description": "Area name -> count of tasks released from it before this re-solve (out of the payload, ADR-0035 decision 3). Never derivable from `preferences` alone \u2014 a release is the absence of a task, not a value in it \u2014 so the frontend, the only side that knows what was released, sends it explicitly. Lets the `reinstated` explanation compare against the plan before the re-solve rather than the payload alone, so a released service the area's own minimum brings back is not reported (it is exactly what release means may happen).",
            "title": "Released",
            "type": "object"
          },
          "scenario": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Scenario"
              },
              {
                "type": "null"
              }
            ],
            "description": "A full problem instance."
          },
          "scenario_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Slug of a curated scenario. Provide this or `scenario`; `scenario` wins when both are set.",
            "title": "Scenario Id"
          },
          "schedule": {
            "$ref": "#/components/schemas/Schedule",
            "description": "The schedule that solve returned."
          }
        },
        "required": [
          "preferences",
          "schedule"
        ],
        "title": "ExplainRequest",
        "type": "object"
      },
      "ExplainResponse": {
        "properties": {
          "report": {
            "$ref": "#/components/schemas/ExplanationReport"
          }
        },
        "required": [
          "report"
        ],
        "title": "ExplainResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RollRequest": {
        "description": "Body for ``POST /api/scenario/advance`` \u2014 the moving-horizon step (ADR-0042).\n\nAdvance \"now\" by ``hours``, given the plan ``tasks`` that were executed. There is no\nexecution simulation: ``tasks`` is just the current plan. ``hours`` out of the\n``1..horizon_hours`` range is a 400.",
        "properties": {
          "hours": {
            "default": 24,
            "description": "How far to advance now, in whole hours.",
            "minimum": 1.0,
            "title": "Hours",
            "type": "integer"
          },
          "scenario": {
            "$ref": "#/components/schemas/Scenario"
          },
          "tasks": {
            "description": "The plan being executed over the elapsed window.",
            "items": {
              "$ref": "#/components/schemas/ScheduledTask"
            },
            "title": "Tasks",
            "type": "array"
          }
        },
        "required": [
          "scenario"
        ],
        "title": "RollRequest",
        "type": "object"
      },
      "RollResponse": {
        "description": "What ``POST /api/scenario/advance`` returns: the rolled scenario (with its derived\nfigures recomputed, like every other scenario response) plus what the roll did.",
        "properties": {
          "carried": {
            "description": "Still-future tasks rebased to the new now, as frozen re-solve preferences.",
            "items": {
              "$ref": "#/components/schemas/PreferredTask"
            },
            "title": "Carried",
            "type": "array"
          },
          "consumed": {
            "description": "New history events from now-past tasks \u2014 one per area serviced in the window.",
            "items": {
              "$ref": "#/components/schemas/ServiceEvent"
            },
            "title": "Consumed",
            "type": "array"
          },
          "derived": {
            "$ref": "#/components/schemas/ScenarioDerived"
          },
          "notes": {
            "description": "Remarks about this roll (areas idle, services straddling now).",
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "scenario": {
            "$ref": "#/components/schemas/Scenario"
          }
        },
        "required": [
          "scenario",
          "derived",
          "consumed",
          "carried",
          "notes"
        ],
        "title": "RollResponse",
        "type": "object"
      },
      "Scenario": {
        "description": "A complete planning problem instance.",
        "properties": {
          "areas": {
            "items": {
              "$ref": "#/components/schemas/Area"
            },
            "title": "Areas",
            "type": "array"
          },
          "base_durations": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/BaseDuration"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Data-driven per-(area, mower) base mowing hours (ADR-0013). When set, the completion table uses these instead of sampling.",
            "title": "Base Durations"
          },
          "duration_seed": {
            "default": 0,
            "description": "Seed for the (otherwise deterministic) mowing-duration sampling. Ignored when base_durations is set.",
            "title": "Duration Seed",
            "type": "integer"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/ServiceEvent"
            },
            "title": "History",
            "type": "array"
          },
          "horizon_hours": {
            "default": 168,
            "exclusiveMinimum": 0.0,
            "title": "Horizon Hours",
            "type": "integer"
          },
          "horizon_start_hour": {
            "default": 13,
            "description": "Hour of the week (0 = Mon 00:00) that t = 0 corresponds to.",
            "maximum": 167.0,
            "minimum": 0.0,
            "title": "Horizon Start Hour",
            "type": "integer"
          },
          "mowers": {
            "items": {
              "$ref": "#/components/schemas/Mower"
            },
            "title": "Mowers",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "areas",
          "mowers",
          "history"
        ],
        "title": "Scenario",
        "type": "object"
      },
      "ScenarioBundle": {
        "properties": {
          "derived": {
            "$ref": "#/components/schemas/ScenarioDerived"
          },
          "scenario": {
            "$ref": "#/components/schemas/Scenario"
          }
        },
        "required": [
          "scenario",
          "derived"
        ],
        "title": "ScenarioBundle",
        "type": "object"
      },
...
      "ScenarioDerived": {
        "description": "Figures computed from the ``Scenario`` \u2014 never stored, recomputed on every edit\n(ADR-0024, ``fleetplanning.derived``). Replaces the old source-derived ``detail``.",
        "properties": {
          "bounds": {
            "additionalProperties": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "type": "integer"
                },
                {
                  "type": "integer"
                }
              ],
              "type": "array"
            },
            "description": "Per-area (min, max) service starts for the week \u2014 the user-owned requirement when set, else the ADR-0017 derivation.",
            "title": "Bounds",
            "type": "object"
          },
          "load_factor": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Demand/capacity proxy; None without size/rate data.",
            "title": "Load Factor"
          }
        },
        "required": [
          "bounds"
        ],
        "title": "ScenarioDerived",
        "type": "object"
      },
...
      "ScenarioTemplateRequest": {
        "description": "Body for ``POST /api/scenario/template`` (ADR-0027). ``name`` seeds the editable\nscenario-name field only \u2014 it is not a slug and nothing is written; Save derives the\nslug from it later.",
        "properties": {
          "name": {
            "default": "new-scenario",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "title": "ScenarioTemplateRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Fleet-Planning Demonstrator API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/explain": {
      "post": {
        "description": "Explain why the edits in `request.preferences` that are missing from\n`request.schedule` were not kept (Iteration 6).\n\nA plain synchronous call, not a job like `/api/solve`: the explain budget (a few\nseconds, capped at 60s by `schemas.ExplainRequest`) bounds it directly, so there is\nnothing to poll or cancel. FastAPI runs this `def` handler in a worker thread, the\nsame as every other handler here \u2014 it does not block the server's event loop, only\nthis one request.",
        "operationId": "post_explain_api_explain_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExplainRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExplainResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Explain"
      }
    },
...
    "/api/scenario/advance": {
      "post": {
        "description": "Advance \"now\" for the moving horizon (ADR-0042): slide the window forward, fold the\nnow-past plan into history, hand back the still-future plan as frozen preferences. The\ncaller then re-solves the returned scenario carrying ``carried``. ``hours`` outside\n``1..horizon_hours`` is a 400.",
        "operationId": "post_scenario_advance_api_scenario_advance_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RollRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RollResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Scenario Advance"
      }
    },
    "/api/scenario/derived": {
      "post": {
        "description": "Recompute the derived figures for an edited scenario (ADR-0024). The editor calls\nthis as the user types; a malformed body is a 422 from request parsing.",
        "operationId": "post_scenario_derived_api_scenario_derived_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Scenario"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioDerived"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Scenario Derived"
      }
    },
    "/api/scenario/template": {
      "post": {
        "description": "A minimal, immediately-solvable scenario for \"New scenario\u2026\" (ADR-0027).\n\nNothing is written: this is a *draft* the editor opens. It reaches disk only via\n``PUT /api/scenario/{id}`` or ``POST /api/scenarios``. 200, not 201 \u2014 no resource\nis created here.",
        "operationId": "post_scenario_template_api_scenario_template_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioTemplateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioBundle"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Scenario Template"
      }
    },
    "/api/scenario/{scenario_id}": {
      "delete": {
        "description": "Remove a scenario from the library (ADR-0026). 404 if the slug is unknown.",
        "operationId": "delete_scenario_api_scenario__scenario_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Scenario"
      },
      "get": {
        "operationId": "get_scenario_api_scenario__scenario_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioBundle"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Scenario"
      },
      "put": {
        "description": "Save the edited scenario over the slug ``scenario_id`` \u2014 create or overwrite\n(ADR-0026). ``scenario.name`` is forced to the slug. A malformed slug is a 400; a\nmalformed scenario is a 422 from request parsing.",
        "operationId": "put_scenario_api_scenario__scenario_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Scenario"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioBundle"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Scenario"
      }
    },
    "/api/scenarios": {
      "get": {
        "operationId": "list_scenarios_api_scenarios_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScenarioListEntry"
                  },
                  "title": "Response List Scenarios Api Scenarios Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Scenarios"
      },
      "post": {
        "description": "Save the edited scenario under a *new* slug (the \"Save as\u2026\" path, ADR-0026). A\nslug that is already taken is a 409 \u2014 use ``PUT`` to overwrite.",
        "operationId": "create_scenario_api_scenarios_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioBundle"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Scenario"
      }
    },
    "/api/solve": {
      "post": {
        "operationId": "post_solve_api_solve_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolveJobStarted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Solve"
      }
    },
    "/api/solve/{job_id}": {
      "get": {
        "operationId": "get_solve_api_solve__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolveJobStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Solve"
      }
    },
    "/api/solve/{job_id}/cancel": {
      "post": {
        "operationId": "post_cancel_solve_api_solve__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolveJobStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Cancel Solve"
      }
    }
  }
}